Refactored mammoth main.py into models and routers
