+++ backends/advanced-backend/.env.template
@@ -16,6 +16,7 @@ NGROK_URL=
 NGROK_AUTHTOKEN=
 HF_TOKEN=
 
+METRICS_COLLECTION_ENABLE=false
 
 
 MEM0_TELEMETRY=
