+++ backends/advanced-backend/src/database.py
@@ -1,0 +1,8 @@
+import os
+from motor.motor_asyncio import AsyncIOMotorClient
+
+MONGODB_URI = os.getenv("MONGODB_URI", "mongodb://mongo:27017")
+mongo_client = AsyncIOMotorClient(MONGODB_URI)
+
+def get_db_client():
+    return mongo_client
