+++ backends/advanced-backend/src/memory/__init__.py
@@ -8,10 +8,12 @@ This module handles all memory-related operations including:
 
 from .memory_service import (
     MemoryService,
-    init_memory_config,
+    # init_memory_config,
     get_memory_service,
     shutdown_memory_service,
 )
+
+from .mem0_client import get_memory_client
 
 __all__ = [
     "MemoryService",
