+++ backends/advanced-backend/src/models/user.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+from typing import Optional
+
+class User(BaseModel):
+    user_id: str
